@@ -1,11 +1,6 @@
 // src/app/interceptors/auth.interceptor.ts
 import { Injectable } from '@angular/core';
-import { 
-  HttpRequest,
-  HttpHandler,
-  HttpEvent,
-  HttpInterceptor
-} from '@angular/common/http';
+import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { AuthService } from '../services/auth.service';
 
